fix: move tempo lanes helpers to `cli::utils`
