+++ crates/cast/src/cmd/batch_send.rs
@@ -15,11 +15,8 @@ use alloy_signer::Signer;
 use clap::Parser;
 use eyre::{Result, eyre};
 use foundry_cli::{
-    opts::{
-        TransactionOpts,
-        tempo_lanes::{maybe_print_resolved_lane, resolve_lane},
-    },
-    utils::{self, LoadConfig},
+    opts::TransactionOpts,
+    utils::{self, LoadConfig, maybe_print_resolved_lane, resolve_lane},
 };
 use foundry_common::provider::ProviderBuilder;
 use std::time::Duration;
